Move Alias/AttributeDicts into utils.
Includes minor fix so they work when deepcopy'd.
